requirements.txt
pytest>=2.9.2
nose==1.3.7
sphinx>=1.4.5
urllib3
socketIO_client
pynacl
requests
websocket-client
easydict
sinon
freezegun
future